Fixing unit testing.
